+++ module/Batch/src/Command/GuildWeeklyCheck/GuildWeeklyCheckFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\GuildWeeklyCheck;
+
+class GuildWeeklyCheckFactory {
+    public function __invoke($controllers): GuildWeeklyCheck
+    {
+        return new GuildWeeklyCheck($controllers->get('faucetdev'));
+    }
+}